Show invalid formula indicator (SHEET-117)
When user types an unparsable formula, cell will have an error indicator.
An error message will be shown in the comment tooltip. If there is a
comment and an invalid formula in a cell, both are shown in the comment
tooltip with a separator line. The error indicator is a red triangle.
The indicator for a comment has been changed to yellow. There is a server
side method to set the error message for invalid formulas to support
localization. Masked reference screenshots of TestBench tests to exclude
the controls on top of the page.

Change-Id: I77dd016f9686c163692b64ecd40c1c2344555bae