Fixes for #461
found that the previous implementation used an implicit null return
value to indicate no conditional formats applied to a cell
(undocumented, bad programming form to use null as a semantic element).  

Had to widen the check to include an empty Set<Integer> as well as
atomically null, since the new implementation always returns a non-null
Set.

I see that all cells with conditional format styles applied are marked
for update and sent to the client on every change.  That is likely fine,
most workbooks don't have that many, but really it should check and
update only on incremental changes.

Also, it didn't notice when conditional format(s) no longer applied to a
cell, and didn't update it to have no formatting in those cases, so
ConditionalFormatter marks all cells where the formatting changed from
the last time it was checked.