Adjust calls to deprecated POI methods removed in POI 4.0.
These changes are all 3.17 compatible, and enable downstream users to
drop in 4.0 builds as needed.

This is a way Spreadsheet users can get the conditional formatting fixes
done in 4.0 without further code changes, once this branch is merged
back and released.

Alternatively, a POI 4.0 build could be lobbied for, and this branch
could move up to that version.