Remove mobile "Fill" overlay, implement context menu for all
Everything but iOS gets the simple ContextMenuEvent handler.

iOS, which doesn't support that, needs a timer to simulate long-press
like Android does natively.

Most of the new code is canceling the timer when a touch is started but
then the user does something else, meaning they don't want the context
menu after all.

Scrolling was the big problem, as it cancelled before the context menu
listener could see it.  Had to cancel the timer from the widget scroll
methods instead.