Allow CSS styling of tabs by name (#700)
By adding the tab name as the tooltip (title attribute) (S)CSS can now
be used to style them by name:

.sheet-tabsheet-tab[title~=TAB_NAME] {
...
}

not all attributes work with this CSS syntax, title is the best I've
found.  Plus, for long tab names, like in the "multiple_sheets.xlsx"
test file, the display name is truncated while the tooltip is the full
name, which is nice.