Move CopyPasteHandler registration to method
To prevent memory leak on the client-side, #660 introduced a method
on the CopyPasteTextBox to remove the handler.
Since the handler registration only occurs on the CopyPasteTextBox
constructor, in some cases when a spreadsheet is detached and
re-attached (if child of VerticalLayout with setSizeFull, for instance),
copy/paste on the spreadsheet stops working.

This PR:
- Removes the registration from CopyPasteTextBox constructor
- Creates a registerHandler method
- Calls the registration from the SheetWidget onAttach method

Fixes #706